@@ -23,7 +23,6 @@
     "react-component"
   ],
   "peerDependencies": {
-    "create-react-class": "^15.6.2",
     "react": "^15.6.2",
     "react-dom": "^15.6.2"
   },
@@ -56,10 +55,9 @@
     "babel-preset-es2015-loose": "^7.0.0",
     "babel-preset-react": "^6.3.13",
     "copy-dir": "^0.3.0",
-    "create-react-class": "^15.6.2",
     "css-loader": "^0.26.0",
     "del": "^1.2.1",
-    "enzyme": "2.8.0",
+    "enzyme": "^2.9.1",
     "eslint": "^1.10.3",
     "eslint-loader": "^1.2.0",
     "eslint-plugin-react": "^3.15.0",
@@ -92,14 +90,12 @@
     "node-dir": "^0.1.12",
     "object-assign": "^2.1.1",
     "open": "0.0.5",
-    "react": "^15.4.2",
+    "react": "^15.6.2",
     "react-addons-perf": "^15.4.2",
-    "react-addons-shallow-compare": "^15.4.2",
-    "react-addons-test-utils": "^15.4.2",
-    "react-addons-transition-group": "^15.4.2",
     "react-dnd-test-backend": "^1.0.2",
     "react-docgen": "^2.8.2",
-    "react-dom": "^15.4.2",
+    "react-dom": "^15.6.2",
+    "react-test-renderer": "^15.6.2",
     "rewire": "^2.1.3",
     "rewire-webpack": "^1.0.0",
     "run-sequence": "^1.2.2",
