@@ -2,7 +2,7 @@ const React = require('react');
 const ReactDOM = require('react-dom');
 const Grid = require('react-data-grid');
 const { editors: { CheckboxEditor } } = Grid;
-const TestUtils = require('react-addons-test-utils');
+const TestUtils = require('react-dom/test-utils');
 const StubComponent = require('../../../../test/StubComponent');
 const mockStateObject = require('./data/MockStateObject');
 const { mount } = require('enzyme');
