@@ -1,6 +1,6 @@
 const React        = require('react');
 const ReactDOM = require('react-dom');
-const TestUtils    = require('react-addons-test-utils');
+const TestUtils    = require('react-dom/test-utils');
 const { Component: ExampleGrid } = require('../scripts/example14-all-features-immutable');
 import { mount } from 'enzyme';
 
