@@ -1,20 +1,16 @@
 const React                 = require('react');
 import PropTypes from 'prop-types';
-const createReactClass = require('create-react-class');
-const ReactDOM = require('react-dom');
 const BaseGrid              = require('./Grid');
-const Row                   = require('./Row');
-const ExcelColumn           = require('./PropTypeShapes/ExcelColumn');
-const KeyboardHandlerMixin  = require('./KeyboardHandlerMixin');
 const CheckboxEditor        = require('./editors/CheckboxEditor');
-const DOMMetrics           = require('./DOMMetrics');
-const ColumnMetricsMixin      = require('./ColumnMetricsMixin');
 const RowUtils = require('./RowUtils');
 const ColumnUtils = require('./ColumnUtils');
 const KeyCodes = require('./KeyCodes');
 const isFunction = require('./utils/isFunction');
 import SelectAll from './formatters/SelectAll';
 import AppConstants from './AppConstants';
+import { DEFINE_SORT } from './cells/headerCells/SortableHeaderCell';
+import { isKeyPrintable, isCtrlKeyHeldDown } from './utils/keyboardUtils';
+const ColumnMetrics = require('./ColumnMetrics');
 require('../../../themes/react-data-grid-core.css');
 require('../../../themes/react-data-grid-checkbox.css');
 
@@ -45,16 +41,16 @@ type RowUpdateEvent = {
   rowIdx: number;
 };
 
-const ReactDataGrid = createReactClass({
-  displayName: 'ReactDataGrid',
+type ColumnMetricsType = {
+  columns: Array<Column>;
+  totalWidth: number;
+  minColumnWidth: number;
+};
 
-  mixins: [
-    ColumnMetricsMixin,
-    DOMMetrics.MetricsComputatorMixin,
-    KeyboardHandlerMixin
-  ],
+class ReactDataGrid extends React.Component {
+  static displayName = 'ReactDataGrid';
 
-  propTypes: {
+  static propTypes = {
     rowHeight: PropTypes.number.isRequired,
     headerRowHeight: PropTypes.number,
     headerFiltersHeight: PropTypes.number,
@@ -73,6 +69,8 @@ const ReactDataGrid = createReactClass({
     getCellActions: PropTypes.func,
     onAddFilter: PropTypes.func,
     onGridSort: PropTypes.func,
+    sortColumn: PropTypes.string,
+    sortDirection: PropTypes.oneOf(Object.keys(DEFINE_SORT)),
     onDragHandleDoubleClick: PropTypes.func,
     onGridRowsUpdated: PropTypes.func,
     onRowSelect: PropTypes.func,
@@ -123,55 +121,207 @@ const ReactDataGrid = createReactClass({
     onAddSubRow: PropTypes.func,
     enableCellAutoFocus: PropTypes.bool,
     onBeforeEdit: PropTypes.func,
-    selectAllRenderer: PropTypes.object
-  },
+    selectAllRenderer: PropTypes.object,
+    minColumnWidth: PropTypes.number,
+    columnEquality: PropTypes.func,
+    onColumnResize: PropTypes.func
+  };
 
-  getDefaultProps(): {enableCellSelect: boolean} {
-    return {
-      enableCellSelect: false,
-      tabIndex: -1,
-      rowHeight: 35,
-      headerFiltersHeight: 45,
-      enableRowSelect: false,
-      minHeight: 350,
-      rowKey: 'id',
-      rowScrollTimeout: 0,
-      scrollToRowIndex: 0,
-      cellNavigationMode: 'none',
-      overScan: {
-        colsStart: 5,
-        colsEnd: 5,
-        rowsStart: 5,
-        rowsEnd: 5
-      },
-      enableCellAutoFocus: true,
-      onBeforeEdit: () => {}
-    };
-  },
+  static defaultProps = {
+    enableCellSelect: false,
+    tabIndex: -1,
+    rowHeight: 35,
+    headerFiltersHeight: 45,
+    enableRowSelect: false,
+    minHeight: 350,
+    rowKey: 'id',
+    rowScrollTimeout: 0,
+    scrollToRowIndex: 0,
+    cellNavigationMode: 'none',
+    overScan: {
+      colsStart: 5,
+      colsEnd: 5,
+      rowsStart: 5,
+      rowsEnd: 5
+    },
+    enableCellAutoFocus: true,
+    onBeforeEdit: () => {},
+    minColumnWidth: 80,
+    columnEquality: ColumnMetrics.sameColumn
+  };
 
-  getInitialState: function(): {selected: SelectedType; copied: ?{idx: number; rowIdx: number}; selectedRows: Array<Row>; expandedRows: Array<Row>; canFilter: boolean; columnFilters: any; sortDirection: ?SortType; sortColumn: ?ExcelColumn; dragged: ?DraggedType;  } {
+  constructor(props, context) {
+    super(props, context);
     let columnMetrics = this.createColumnMetrics();
     let initialState = {columnMetrics, selectedRows: [], copied: null, expandedRows: [], canFilter: false, columnFilters: {}, sortDirection: null, sortColumn: null, dragged: null, scrollOffset: 0, lastRowIdxUiSelected: -1};
-    if (this.props.enableCellSelect) {
+    if (props.enableCellSelect) {
       initialState.selected = {rowIdx: 0, idx: 0};
     } else {
       initialState.selected = {rowIdx: -1, idx: -1};
     }
-    return initialState;
-  },
 
-  hasSelectedCellChanged: function(selected: SelectedType) {
+    if (this.props.sortColumn && this.props.sortDirection) {
+      initialState.sortColumn = this.props.sortColumn;
+      initialState.sortDirection = this.props.sortDirection;
+    }
+
+    this.state = initialState;
+  }
+
+  componentWillMount() {
+    this._mounted = true;
+  }
+
+  componentDidMount() {
+    if (window.addEventListener) {
+      window.addEventListener('resize', this.metricsUpdated);
+    } else {
+      window.attachEvent('resize', this.metricsUpdated);
+    }
+    this.metricsUpdated();
+  }
+
+  componentWillUnmount() {
+    this._mounted = false;
+    window.removeEventListener('resize', this.metricsUpdated);
+  }
+
+  componentWillReceiveProps(nextProps) {
+    if (nextProps.columns) {
+      if (!ColumnMetrics.sameColumns(this.props.columns, nextProps.columns, this.props.columnEquality) ||
+          nextProps.minWidth !== this.props.minWidth) {
+        let columnMetrics = this.createColumnMetrics(nextProps);
+        this.setState({columnMetrics: columnMetrics});
+      }
+    }
+  }
+
+  gridWidth = (): number => {
+    return this.grid ? this.grid.parentElement.offsetWidth : 0;
+  };
+
+  getTotalWidth = () => {
+    let totalWidth = 0;
+    if (this._mounted) {
+      totalWidth = this.gridWidth();
+    } else {
+      totalWidth = ColumnUtils.getSize(this.props.columns) * this.props.minColumnWidth;
+    }
+    return totalWidth;
+  };
+
+  getColumnMetricsType = (metrics: ColumnMetricsType): { columns: ColumnMetricsType } => {
+    let totalWidth = metrics.totalWidth || this.getTotalWidth();
+    let currentMetrics = {
+      columns: metrics.columns,
+      totalWidth: totalWidth,
+      minColumnWidth: metrics.minColumnWidth
+    };
+    let updatedMetrics = ColumnMetrics.recalculate(currentMetrics);
+    return updatedMetrics;
+  };
+
+  getColumn = (idx) => {
+    let columns = this.state.columnMetrics.columns;
+    if (Array.isArray(columns)) {
+      return columns[idx];
+    }else if (typeof Immutable !== 'undefined') {
+      return columns.get(idx);
+    }
+  };
+
+  getSize = () => {
+    let columns = this.state.columnMetrics.columns;
+    if (Array.isArray(columns)) {
+      return columns.length;
+    }else if (typeof Immutable !== 'undefined') {
+      return columns.size;
+    }
+  };
+
+  metricsUpdated = () => {
+    let columnMetrics = this.createColumnMetrics();
+    this.setState({columnMetrics});
+  };
+
+  createColumnMetrics = (props = this.props) => {
+    let gridColumns = this.setupGridColumns(props);
+    return this.getColumnMetricsType({
+      columns: gridColumns,
+      minColumnWidth: this.props.minColumnWidth,
+      totalWidth: props.minWidth
+    });
+  };
+
+  onColumnResize = (index: number, width: number) => {
+    let columnMetrics = ColumnMetrics.resizeColumn(this.state.columnMetrics, index, width);
+    this.setState({columnMetrics});
+    if (this.props.onColumnResize) {
+      this.props.onColumnResize(index, width);
+    }
+  };
+
+  onKeyDown = (e: SyntheticKeyboardEvent) => {
+    if (isCtrlKeyHeldDown(e)) {
+      this.checkAndCall('onPressKeyWithCtrl', e);
+    } else if (this.isKeyExplicitlyHandled(e.key)) {
+      // break up individual keyPress events to have their own specific callbacks
+      let callBack = 'onPress' + e.key;
+      this.checkAndCall(callBack, e);
+    } else if (isKeyPrintable(e.keyCode)) {
+      this.checkAndCall('onPressChar', e);
+    }
+
+    // Track which keys are currently down for shift clicking etc
+    this._keysDown = this._keysDown || {};
+    this._keysDown[e.keyCode] = true;
+    if (isFunction(this.props.onGridKeyDown)) {
+      this.props.onGridKeyDown(e);
+    }
+  };
+
+  onKeyUp = (e) => {
+    // Track which keys are currently down for shift clicking etc
+    this._keysDown = this._keysDown || {};
+    delete this._keysDown[e.keyCode];
+
+    if (isFunction(this.props.onGridKeyUp)) {
+      this.props.onGridKeyUp(e);
+    }
+  };
+
+  isKeyDown = (keyCode) => {
+    if (!this._keysDown) return false;
+    return keyCode in this._keysDown;
+  };
+
+  isSingleKeyDown = (keyCode) => {
+    if (!this._keysDown) return false;
+    return keyCode in this._keysDown && Object.keys(this._keysDown).length === 1;
+  };
+
+  checkAndCall = (methodName: string, args: any) => {
+    if (typeof this[methodName] === 'function') {
+      this[methodName](args);
+    }
+  };
+
+  isKeyExplicitlyHandled = (key: string): boolean => {
+    return typeof this['onPress' + key] === 'function';
+  };
+
+  hasSelectedCellChanged = (selected: SelectedType) => {
     let previouslySelected = Object.assign({}, this.state.selected);
     return previouslySelected.rowIdx !== selected.rowIdx || previouslySelected.idx !== selected.idx || previouslySelected.active === false;
-  },
+  };
 
-  onContextMenuHide: function() {
+  onContextMenuHide = () => {
     document.removeEventListener('click', this.onContextMenuHide);
     let newSelected = Object.assign({}, this.state.selected, {contextMenuDisplayed: false});
     this.setState({selected: newSelected});
-  },
+  };
 
-  onColumnEvent: function(ev :SyntheticEvent, columnEvent: ColumnEvent) {
+  onColumnEvent = (ev :SyntheticEvent, columnEvent: ColumnEvent) => {
     let {idx, name} = columnEvent;
 
     if (name && typeof idx !== 'undefined') {
@@ -188,9 +338,9 @@ const ReactDataGrid = createReactClass({
         column.events[name](ev, eventArgs);
       }
     }
-  },
+  };
 
-  onSelect: function(selected: SelectedType) {
+  onSelect = (selected: SelectedType) => {
     if (this.state.selected.rowIdx !== selected.rowIdx
       || this.state.selected.idx !== selected.idx
       || this.state.selected.active === false) {
@@ -211,9 +361,9 @@ const ReactDataGrid = createReactClass({
         this.setState({selected: { idx, rowIdx }});
       }
     }
-  },
+  };
 
-  onCellClick: function(cell: SelectedType, e: SyntheticEvent) {
+  onCellClick = (cell: SelectedType, e: SyntheticEvent) => {
     this.onSelect({rowIdx: cell.rowIdx, idx: cell.idx});
 
     if (this.props.onRowClick && typeof this.props.onRowClick === 'function') {
@@ -223,16 +373,20 @@ const ReactDataGrid = createReactClass({
     if (e) {
       e.stopPropagation();
     }
-  },
+  };
 
-  onCellContextMenu: function(cell: SelectedType) {
+  onCellFocus = (cell: SelectedType) => {
+    this.onSelect(cell);
+  };
+
+  onCellContextMenu = (cell: SelectedType) => {
     this.onSelect({rowIdx: cell.rowIdx, idx: cell.idx, contextMenuDisplayed: this.props.contextMenu});
     if (this.props.contextMenu) {
       document.addEventListener('click', this.onContextMenuHide);
     }
-  },
+  };
 
-  onCellDoubleClick: function(cell: SelectedType, e: SyntheticEvent) {
+  onCellDoubleClick = (cell: SelectedType, e: SyntheticEvent) => {
     this.onSelect({rowIdx: cell.rowIdx, idx: cell.idx});
     if (this.props.onRowDoubleClick && typeof this.props.onRowDoubleClick === 'function') {
       this.props.onRowDoubleClick(cell.rowIdx, this.props.rowGetter(cell.rowIdx), this.getColumn(cell.idx));
@@ -241,51 +395,51 @@ const ReactDataGrid = createReactClass({
     if (e) {
       e.stopPropagation();
     }
-  },
+  };
 
-  onPressArrowUp(e: SyntheticEvent) {
+  onPressArrowUp = (e: SyntheticEvent) => {
     this.moveSelectedCell(e, -1, 0);
-  },
+  };
 
-  onPressArrowDown(e: SyntheticEvent) {
+  onPressArrowDown = (e: SyntheticEvent) => {
     this.moveSelectedCell(e, 1, 0);
-  },
+  };
 
-  onPressArrowLeft(e: SyntheticEvent) {
+  onPressArrowLeft = (e: SyntheticEvent) => {
     this.moveSelectedCell(e, 0, -1);
-  },
+  };
 
-  onPressArrowRight(e: SyntheticEvent) {
+  onPressArrowRight = (e: SyntheticEvent) => {
     this.moveSelectedCell(e, 0, 1);
-  },
+  };
 
-  isFocusedOnCell() {
+  isFocusedOnCell = () => {
     return document.activeElement && document.activeElement.classList &&
       document.activeElement.classList.contains('react-grid-Cell');
-  },
+  };
 
-  isFocusedOnTable() {
+  isFocusedOnTable = () => {
     const domNode = this.getDataGridDOMNode();
     return domNode && domNode.contains(document.activeElement);
-  },
+  };
 
-  exitGrid(oldSelectedCell, newSelectedValue) {
+  exitGrid = (oldSelectedCell, newSelectedValue) => {
     this.setState({ selected: newSelectedValue },
       () => {
         if (typeof this.props.onCellDeSelected === 'function') {
           this.props.onCellDeSelected(oldSelectedCell);
         }});
-  },
+  };
 
-  enterGrid(newSelectedValue) {
+  enterGrid = (newSelectedValue) => {
     this.setState({ selected: newSelectedValue },
       () => {
         if (typeof this.props.onCellSelected === 'function') {
           this.props.onCellSelected(newSelectedValue);
         }});
-  },
+  };
 
-  onPressTab(e: SyntheticEvent) {
+  onPressTab = (e: SyntheticEvent) => {
     // Scenario 0a: When there are no rows in the grid, pressing tab needs to allow the browser to handle it
     if (this.props.rowsCount === 0) {
       return;
@@ -341,32 +495,32 @@ const ReactDataGrid = createReactClass({
       return;
     }
     this.moveSelectedCell(e, 0, e.shiftKey ? -1 : 1);
-  },
+  };
 
-  onPressEnter(e: SyntheticKeyboardEvent) {
+  onPressEnter = (e: SyntheticKeyboardEvent) => {
     this.setActive(e.key);
-  },
+  };
 
-  onPressDelete(e: SyntheticKeyboardEvent) {
+  onPressDelete = (e: SyntheticKeyboardEvent) => {
     this.setActive(e.key);
-  },
+  };
 
-  onPressEscape(e: SyntheticKeyboardEvent) {
+  onPressEscape = (e: SyntheticKeyboardEvent) => {
     this.setInactive(e.key);
     this.handleCancelCopy();
-  },
+  };
 
-  onPressBackspace(e: SyntheticKeyboardEvent) {
+  onPressBackspace = (e: SyntheticKeyboardEvent) => {
     this.setActive(e.key);
-  },
+  };
 
-  onPressChar(e: SyntheticKeyboardEvent) {
-    if (this.isKeyPrintable(e.keyCode)) {
+  onPressChar = (e: SyntheticKeyboardEvent) => {
+    if (isKeyPrintable(e.keyCode)) {
       this.setActive(e.keyCode);
     }
-  },
+  };
 
-  onPressKeyWithCtrl(e: SyntheticKeyboardEvent) {
+  onPressKeyWithCtrl = (e: SyntheticKeyboardEvent) => {
     let keys = {
       KeyCode_c: 99,
       KeyCode_C: 67,
@@ -388,9 +542,9 @@ const ReactDataGrid = createReactClass({
         this.handlePaste();
       }
     }
-  },
+  };
 
-  onGridRowsUpdated(cellKey, fromRow, toRow, updated, action, originRow) {
+  onGridRowsUpdated = (cellKey, fromRow, toRow, updated, action, originRow) => {
     let rowIds = [];
 
     for (let i = fromRow; i <= toRow; i++) {
@@ -401,9 +555,9 @@ const ReactDataGrid = createReactClass({
     let fromRowId = fromRowData[this.props.rowKey];
     let toRowId = this.props.rowGetter(toRow)[this.props.rowKey];
     this.props.onGridRowsUpdated({cellKey, fromRow, toRow, fromRowId, toRowId, rowIds, updated, action, fromRowData});
-  },
+  };
 
-  onCellCommit(commit: RowUpdateEvent) {
+  onCellCommit = (commit: RowUpdateEvent) => {
     let selected = Object.assign({}, this.state.selected);
     selected.active = false;
     let expandedRows = this.state.expandedRows;
@@ -421,9 +575,9 @@ const ReactDataGrid = createReactClass({
     if (this.props.onGridRowsUpdated) {
       this.onGridRowsUpdated(commit.cellKey, targetRow, targetRow, commit.updated, AppConstants.UpdateActions.CELL_UPDATE);
     }
-  },
+  };
 
-  onDragStart(e: SyntheticEvent) {
+  onDragStart = (e: SyntheticEvent) => {
     let idx = this.state.selected.idx;
     // To prevent dragging down/up when reordering rows.
     const isViewportDragging = e && e.target && e.target.className;
@@ -439,9 +593,9 @@ const ReactDataGrid = createReactClass({
         }
       }
     }
-  },
+  };
 
-  onToggleFilter() {
+  onToggleFilter = () => {
     // setState() does not immediately mutate this.state but creates a pending state transition.
     // Therefore if you want to do something after the state change occurs, pass it in as a callback function.
     this.setState({ canFilter: !this.state.canFilter }, () => {
@@ -449,9 +603,9 @@ const ReactDataGrid = createReactClass({
         this.props.onClearFilters();
       }
     });
-  },
+  };
 
-  onDragHandleDoubleClick(e) {
+  onDragHandleDoubleClick = (e) => {
     if (this.props.onDragHandleDoubleClick) {
       this.props.onDragHandleDoubleClick(e);
     }
@@ -460,35 +614,35 @@ const ReactDataGrid = createReactClass({
       let cellKey = this.getColumn(e.idx).key;
       this.onGridRowsUpdated(cellKey, e.rowIdx, this.props.rowsCount - 1, {[cellKey]: e.rowData[cellKey]}, AppConstants.UpdateActions.COLUMN_FILL);
     }
-  },
+  };
 
-  onCellExpand(args) {
+  onCellExpand = (args) => {
     if (this.props.onCellExpand) {
       this.props.onCellExpand(args);
     }
-  },
+  };
 
-  onRowExpandToggle(args) {
+  onRowExpandToggle = (args) => {
     if (typeof this.props.onRowExpandToggle === 'function') {
       this.props.onRowExpandToggle(args);
     }
-  },
+  };
 
-  isCellWithinBounds({idx, rowIdx}) {
+  isCellWithinBounds = ({idx, rowIdx}) => {
     return idx >= 0
       && rowIdx >= 0
       && idx < ColumnUtils.getSize(this.state.columnMetrics.columns)
       && rowIdx < this.props.rowsCount;
-  },
+  };
 
-  handleDragStart(dragged: DraggedType) {
+  handleDragStart = (dragged: DraggedType) => {
     if (!this.dragEnabled()) { return; }
     if (this.isCellWithinBounds(dragged)) {
       this.setState({ dragged: dragged });
     }
-  },
+  };
 
-  handleDragEnd() {
+  handleDragEnd = () => {
     if (!this.dragEnabled()) { return; }
     const { selected, dragged } = this.state;
     const column = this.getColumn(this.state.selected.idx);
@@ -504,21 +658,21 @@ const ReactDataGrid = createReactClass({
       }
     }
     this.setState({dragged: {complete: true}});
-  },
+  };
 
-  handleDragEnter(row: any) {
+  handleDragEnter = (row: any) => {
     if (!this.dragEnabled() || this.state.dragged == null) { return; }
     let dragged = this.state.dragged;
     dragged.overRowIdx = row;
     this.setState({dragged: dragged});
-  },
+  };
 
-  handleTerminateDrag() {
+  handleTerminateDrag = () => {
     if (!this.dragEnabled()) { return; }
     this.setState({ dragged: null });
-  },
+  };
 
-  handlePaste() {
+  handlePaste = () => {
     if (!this.copyPasteEnabled() || !(this.state.copied)) { return; }
     let selected = this.state.selected;
     let cellKey = this.getColumn(this.state.selected.idx).key;
@@ -533,27 +687,27 @@ const ReactDataGrid = createReactClass({
     if (this.props.onGridRowsUpdated) {
       this.onGridRowsUpdated(cellKey, toRow, toRow, {[cellKey]: textToCopy}, AppConstants.UpdateActions.COPY_PASTE, fromRow);
     }
-  },
+  };
 
-  handleCancelCopy() {
+  handleCancelCopy = () => {
     this.setState({copied: null});
-  },
+  };
 
-  handleCopy(args: {value: string}) {
+  handleCopy = (args: {value: string}) => {
     if (!this.copyPasteEnabled()) { return; }
     let textToCopy = args.value;
     let selected = this.state.selected;
     let copied = {idx: selected.idx, rowIdx: selected.rowIdx};
     this.setState({textToCopy: textToCopy, copied: copied});
-  },
+  };
 
-  handleSort: function(columnKey: string, direction: SortType) {
+  handleSort = (columnKey: string, direction: SortType) => {
     this.setState({sortDirection: direction, sortColumn: columnKey}, function() {
       this.props.onGridSort(columnKey, direction);
     });
-  },
+  };
 
-  getSelectedRow(rows, key) {
+  getSelectedRow = (rows, key) => {
     let selectedRow = rows.filter(r => {
       if (r[this.props.rowKey] === key) {
         return true;
@@ -563,14 +717,14 @@ const ReactDataGrid = createReactClass({
     if (selectedRow.length > 0) {
       return selectedRow[0];
     }
-  },
+  };
 
-  useNewRowSelection() {
+  useNewRowSelection = () => {
     return this.props.rowSelection && this.props.rowSelection.selectBy;
-  },
+  };
 
   // return false if not a shift select so can be handled as normal row selection
-  handleShiftSelect(rowIdx) {
+  handleShiftSelect = (rowIdx) => {
     if (this.state.lastRowIdxUiSelected > -1 && this.isSingleKeyDown(KeyCodes.Shift)) {
       let {keys, indexes, isSelectedKey} = this.props.rowSelection.selectBy;
       let isPreviouslySelected = RowUtils.isRowSelected(keys, indexes, isSelectedKey, this.props.rowGetter(rowIdx), rowIdx);
@@ -613,9 +767,9 @@ const ReactDataGrid = createReactClass({
     }
 
     return false;
-  },
+  };
 
-  handleNewRowSelect(rowIdx, rowData) {
+  handleNewRowSelect = (rowIdx, rowData) => {
     if (this.selectAllCheckbox && this.selectAllCheckbox.checked === true) {
       this.selectAllCheckbox.checked = false;
     }
@@ -630,11 +784,11 @@ const ReactDataGrid = createReactClass({
     } else if (!isPreviouslySelected && typeof this.props.rowSelection.onRowsSelected === 'function') {
       this.props.rowSelection.onRowsSelected([{rowIdx, row: rowData}]);
     }
-  },
+  };
 
   // columnKey not used here as this function will select the whole row,
   // but needed to match the function signature in the CheckboxEditor
-  handleRowSelect(rowIdx: number, columnKey: string, rowData, e: Event) {
+  handleRowSelect = (rowIdx: number, columnKey: string, rowData, e: Event) => {
     e.stopPropagation();
 
     if (this.useNewRowSelection()) {
@@ -659,9 +813,9 @@ const ReactDataGrid = createReactClass({
         this.props.onRowSelect(selectedRows.filter(r => r.isSelected === true));
       }
     }
-  },
+  };
 
-  handleCheckboxChange: function(e: SyntheticEvent) {
+  handleCheckboxChange = (e: SyntheticEvent) => {
     let allRowsSelected;
     if (e.currentTarget instanceof HTMLInputElement && e.currentTarget.checked === true) {
       allRowsSelected = true;
@@ -707,24 +861,26 @@ const ReactDataGrid = createReactClass({
         this.props.onRowSelect(selectedRows.filter(r => r.isSelected === true));
       }
     }
-  },
+  };
 
-  getScrollOffSet() {
+  getScrollOffSet = () => {
     let scrollOffset = 0;
-    let canvas = ReactDOM.findDOMNode(this).querySelector('.react-grid-Canvas');
-    if (canvas) {
-      scrollOffset = canvas.offsetWidth - canvas.clientWidth;
+    if (this.grid) {
+      let canvas = this.grid.querySelector('.react-grid-Canvas');
+      if (canvas) {
+        scrollOffset = canvas.offsetWidth - canvas.clientWidth;
+      }
     }
     this.setState({scrollOffset: scrollOffset});
-  },
+  };
 
-  getRowOffsetHeight(): number {
+  getRowOffsetHeight = (): number => {
     let offsetHeight = 0;
     this.getHeaderRows().forEach((row) => offsetHeight += parseFloat(row.height, 10) );
     return offsetHeight;
-  },
+  };
 
-  getHeaderRows(): Array<{ref: Function; height: number;}> {
+  getHeaderRows = (): Array<{ref: Function; height: number;}> => {
     let rows = [{ ref: (node) => this.row = node, height: this.props.headerRowHeight || this.props.rowHeight, rowType: 'header' }];
     if (this.state.canFilter === true) {
       rows.push({
@@ -736,40 +892,41 @@ const ReactDataGrid = createReactClass({
       });
     }
     return rows;
-  },
+  };
 
-  getInitialSelectedRows: function() {
+  getInitialSelectedRows = () => {
     let selectedRows = [];
     for (let i = 0; i < this.props.rowsCount; i++) {
       selectedRows.push(false);
     }
     return selectedRows;
-  },
+  };
 
-  getRowSelectionProps() {
+  getRowSelectionProps = () => {
     if (this.props.rowSelection) {
       return this.props.rowSelection.selectBy;
     }
 
     return null;
-  },
+  };
 
-  getSelectedRows() {
+  getSelectedRows = () => {
     if (this.props.rowSelection) {
       return null;
     }
 
     return this.state.selectedRows.filter(r => r.isSelected === true);
-  },
+  };
 
-  getSelectedValue(): string {
+  getSelectedValue = (): string => {
     let rowIdx = this.state.selected.rowIdx;
     let idx = this.state.selected.idx;
     let cellKey = this.getColumn(idx).key;
     let row = this.props.rowGetter(rowIdx);
     return RowUtils.get(row, cellKey);
-  },
-  canExitGrid(e: SyntheticEvent): boolean {
+  };
+
+  canExitGrid = (e: SyntheticEvent): boolean => {
     // When the cellNavigationMode is 'none', you can exit the grid if you're at the start or end of the row
     // When the cellNavigationMode is 'changeRow', you can exit the grid if you're at the first or last cell of the grid
     // When the cellNavigationMode is 'loopOverRow', there is no logical exit point so you can't exit the grid
@@ -801,9 +958,9 @@ const ReactDataGrid = createReactClass({
       }
     }
     return false;
-  },
+  };
 
-  moveSelectedCell(e: SyntheticEvent, rowDelta: number, cellDelta: number) {
+  moveSelectedCell = (e: SyntheticEvent, rowDelta: number, cellDelta: number) => {
     // we need to prevent default as we control grid scroll
     // otherwise it moves every time you left/right which is janky
     e.preventDefault();
@@ -818,21 +975,18 @@ const ReactDataGrid = createReactClass({
     }
     this.scrollToColumn(idx);
     this.onSelect({ idx: idx, rowIdx: rowIdx });
-  },
+  };
 
-  getNbrColumns() {
+  getNbrColumns = () => {
     const {columns, enableRowSelect} = this.props;
     return enableRowSelect ? columns.length + 1 : columns.length;
-  },
+  };
 
-  getDataGridDOMNode() {
-    if (!this._gridNode) {
-      this._gridNode = ReactDOM.findDOMNode(this);
-    }
-    return this._gridNode;
-  },
+  getDataGridDOMNode = () => {
+    return this.grid;
+  };
 
-  calculateNextSelectionPosition(cellNavigationMode: string, cellDelta: number, rowDelta: number) {
+  calculateNextSelectionPosition = (cellNavigationMode: string, cellDelta: number, rowDelta: number) => {
     let _rowDelta = rowDelta;
     let idx = this.state.selected.idx + cellDelta;
     const nbrColumns = this.getNbrColumns();
@@ -857,25 +1011,25 @@ const ReactDataGrid = createReactClass({
     }
     let rowIdx = this.state.selected.rowIdx + _rowDelta;
     return {idx, rowIdx};
-  },
+  };
 
-  isAtLastCellInRow(nbrColumns) {
+  isAtLastCellInRow = (nbrColumns) => {
     return this.state.selected.idx === nbrColumns - 1;
-  },
+  };
 
-  isAtLastRow() {
+  isAtLastRow = () => {
     return this.state.selected.rowIdx === this.props.rowsCount - 1;
-  },
+  };
 
-  isAtFirstCellInRow() {
+  isAtFirstCellInRow = () => {
     return this.state.selected.idx === 0;
-  },
+  };
 
-  isAtFirstRow() {
+  isAtFirstRow = () => {
     return this.state.selected.rowIdx === 0;
-  },
+  };
 
-  openCellEditor(rowIdx, idx) {
+  openCellEditor = (rowIdx, idx) => {
     let row = this.props.rowGetter(rowIdx);
     let col = this.getColumn(idx);
 
@@ -891,47 +1045,49 @@ const ReactDataGrid = createReactClass({
     } else {
       this.setActive('Enter');
     }
-  },
+  };
 
-  scrollToColumn(colIdx) {
-    let canvas = ReactDOM.findDOMNode(this).querySelector('.react-grid-Canvas');
-    if (canvas) {
-      let left = 0;
-      let locked = 0;
+  scrollToColumn = (colIdx) => {
+    if (this.grid) {
+      let canvas = this.grid.querySelector('.react-grid-Canvas');
+      if (canvas) {
+        let left = 0;
+        let locked = 0;
 
-      for (let i = 0; i < colIdx; i++) {
-        let column = this.getColumn(i);
-        if (column) {
-          if (column.width) {
-            left += column.width;
-          }
-          if (column.locked) {
-            locked += column.width;
+        for (let i = 0; i < colIdx; i++) {
+          let column = this.getColumn(i);
+          if (column) {
+            if (column.width) {
+              left += column.width;
+            }
+            if (column.locked) {
+              locked += column.width;
+            }
           }
         }
-      }
 
-      let selectedColumn = this.getColumn(colIdx);
-      if (selectedColumn) {
-        let scrollLeft = left - locked - canvas.scrollLeft;
-        let scrollRight =  left + selectedColumn.width - canvas.scrollLeft;
+        let selectedColumn = this.getColumn(colIdx);
+        if (selectedColumn) {
+          let scrollLeft = left - locked - canvas.scrollLeft;
+          let scrollRight =  left + selectedColumn.width - canvas.scrollLeft;
 
-        if (scrollLeft < 0) {
-          canvas.scrollLeft += scrollLeft;
-        } else if (scrollRight > canvas.clientWidth) {
-          let scrollAmount =  scrollRight - canvas.clientWidth;
-          canvas.scrollLeft += scrollAmount;
+          if (scrollLeft < 0) {
+            canvas.scrollLeft += scrollLeft;
+          } else if (scrollRight > canvas.clientWidth) {
+            let scrollAmount =  scrollRight - canvas.clientWidth;
+            canvas.scrollLeft += scrollAmount;
+          }
         }
       }
     }
-  },
+  };
 
-  deselect() {
+  deselect = () => {
     const selected = {rowIdx: -1, idx: -1};
     this.setState({selected});
-  },
+  };
 
-  setActive(keyPressed: string) {
+  setActive = (keyPressed: string) => {
     let rowIdx = this.state.selected.rowIdx;
     let row = this.props.rowGetter(rowIdx);
 
@@ -955,9 +1111,9 @@ const ReactDataGrid = createReactClass({
         this.handleCancelCopy();
       }
     }
-  },
+  };
 
-  setInactive() {
+  setInactive = () => {
     let rowIdx = this.state.selected.rowIdx;
     let row = this.props.rowGetter(rowIdx);
 
@@ -968,13 +1124,13 @@ const ReactDataGrid = createReactClass({
       let selected = Object.assign({}, this.state.selected, {idx: idx, rowIdx: rowIdx, active: false});
       this.setState({selected: selected});
     }
-  },
+  };
 
-  isActive(): boolean {
+  isActive = (): boolean => {
     return this.state.selected.active === true;
-  },
+  };
 
-  setupGridColumns: function(props = this.props): Array<any> {
+  setupGridColumns = (props = this.props): Array<any> => {
     const { columns } = props;
     if (this._cachedColumns === columns) {
       return this._cachedComputedColumns;
@@ -1007,17 +1163,17 @@ const ReactDataGrid = createReactClass({
     this._cachedComputedColumns = cols;
 
     return this._cachedComputedColumns;
-  },
+  };
 
-  copyPasteEnabled: function(): boolean {
+  copyPasteEnabled = (): boolean => {
     return this.props.onCellCopyPaste !== null;
-  },
+  };
 
-  dragEnabled: function(): boolean {
+  dragEnabled = (): boolean => {
     return this.props.onGridRowsUpdated !== undefined || this.props.onCellsDragged !== undefined;
-  },
+  };
 
-  renderToolbar(): ReactElement {
+  renderToolbar = (): ReactElement => {
     let Toolbar = this.props.toolbar;
     let toolBarProps =  {columns: this.props.columns, onToggleFilter: this.onToggleFilter, numberOfRows: this.props.rowsCount};
     if (React.isValidElement(Toolbar)) {
@@ -1025,7 +1181,7 @@ const ReactDataGrid = createReactClass({
     } else if (isFunction(Toolbar)) {
       return <Toolbar {...toolBarProps}/>;
     }
-  },
+  };
 
   render() {
     let cellMetaData = {
@@ -1034,6 +1190,7 @@ const ReactDataGrid = createReactClass({
       dragged: this.state.dragged,
       hoveredRowIdx: this.state.hoveredRowIdx,
       onCellClick: this.onCellClick,
+      onCellFocus: this.onCellFocus,
       onCellContextMenu: this.onCellContextMenu,
       onCellDoubleClick: this.onCellDoubleClick,
       onCommit: this.onCellCommit,
@@ -1058,7 +1215,7 @@ const ReactDataGrid = createReactClass({
     };
 
     let toolbar = this.renderToolbar();
-    let containerWidth = this.props.minWidth || this.DOMMetrics.gridWidth();
+    let containerWidth = this.props.minWidth || this.gridWidth();
     let gridWidth = containerWidth - this.state.scrollOffset;
 
     // depending on the current lifecycle stage, gridWidth() may not initialize correctly
@@ -1071,7 +1228,8 @@ const ReactDataGrid = createReactClass({
       gridWidth = '100%';
     }
     return (
-      <div className="react-grid-Container" style={{width: containerWidth}}>
+      <div className="react-grid-Container" style={{width: containerWidth}}
+        ref={(node) => { this.grid = node; }}>
         {toolbar}
         <div className="react-grid-Main">
           <BaseGrid
@@ -1108,7 +1266,7 @@ const ReactDataGrid = createReactClass({
         </div>
       );
   }
-});
+}
 
 
 module.exports = ReactDataGrid;
