@@ -2,7 +2,7 @@ const React         = require('react');
 const ReactDOM = require('react-dom');
 const rewire        = require('rewire');
 const Grid          = rewire('../Grid');
-const TestUtils     = require('react-addons-test-utils');
+const TestUtils     = require('react-dom/test-utils');
 const helpers       = require('../helpers/test/GridPropHelpers');
 const rewireModule = require('../../../../test/rewireModule');
 import { shallow } from 'enzyme';
