@@ -1,7 +1,7 @@
 const React         = require('react');
 const rewire        = require('rewire');
 const HeaderCell    = rewire('../HeaderCell');
-const TestUtils     = require('react-addons-test-utils');
+const TestUtils     = require('react-dom/test-utils');
 const rewireModule  = require('../../../../test/rewireModule');
 const StubComponent = require('../../../../test/StubComponent');
 import SortableHeaderCell from '../cells/headerCells/SortableHeaderCell';
