@@ -1,6 +1,6 @@
 const React         = require('react');
 const rewire        = require('rewire');
-const TestUtils     = require('react-addons-test-utils');
+const TestUtils     = require('react-dom/test-utils');
 const rewireModule  = require('../../../../test/rewireModule');
 const StubComponent = require('../../../../test/StubComponent');
 const helpers       = require('../helpers/test/GridPropHelpers');
