@@ -1,5 +1,5 @@
 import React from 'react';
-import TestUtils from 'react-addons-test-utils';
+import TestUtils from 'react-dom/test-utils';
 import Row from '../Row';
 import { shallow } from 'enzyme';
 import * as helpers from '../helpers/test/GridPropHelpers';
