@@ -1,6 +1,6 @@
 const React          = require('react');
 const ReactDOM = require('react-dom');
-const TestUtils      = require('react-addons-test-utils');
+const TestUtils      = require('react-dom/test-utils');
 const CheckboxEditor = require('../CheckboxEditor');
 const { mount } = require('enzyme');
 
