@@ -3,7 +3,7 @@ const React            = require('react');
 const ReactDOM         = require('react-dom');
 const rewire           = require('rewire');
 const EditorContainer  = rewire('../EditorContainer.js');
-const TestUtils        = require('react-addons-test-utils');
+const TestUtils        = require('react-dom/test-utils');
 const SimpleTextEditor = require('../SimpleTextEditor');
 const EditorBase       = require('../EditorBase');
 import { shallow } from 'enzyme';
