@@ -2,7 +2,7 @@
 var gridHelpers = require('../browser/helpers.js');
 var React = require('react');
 var ReactDOM = require('react-dom');
-var ReactTests = require('react-addons-test-utils');
+var ReactTests = require('react-dom/test-utils');
 var  Grid = require('../../src/Grid');
 var  Viewport = require('../../src/Viewport');
 var  Canvas = require('../../src/Canvas');
